requirements.txt
Django==3.2.4
django_filter==21.1
djangorestframework==3.13.1
drf_yasg==1.20.0
PyMySQL==1.0.2
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
black==22.3.0
